refactor(tests): add boilerplate test files
